Reorg the shape and text style type defs
